[refactor] 포트폴리오용 버튼 관련 스타일 리펙터링
버튼 이름 reverse-> empty로
reverse 버튼 한개만 남겨두고 나머지는 포트폴리오 전용
스타일에서 변형해서 적용하기'
