--- client/src/basics/ReverseButton/Green.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import ReverseButton from '.';
-
-const LongReverseButton = styled(ReverseButton)`
-    width: 5rem;
-`;
-
-export default LongReverseButton;
