--- client/src/basics/ReverseButton/Red.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-import ReverseButton from '.';
-
-const LongReverseButton = styled(ReverseButton)`
-    width: 5rem;
-    color: red;
-    border: 2px red solid ;
-`;
-
-export default LongReverseButton;
